requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.24.1
supabase==2.0.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
psutil==5.9.6